clipbuddy/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
distro==1.9.0
Django==4.2.13
django-environ==0.11.2
exceptiongroup==1.2.1
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.292
langchain-community==0.2.0
langchain-core==0.2.0
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langsmith==0.0.92
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.10.0
numpy==1.26.4
openai==0.28.0
orjson==3.10.3
packaging==23.2
pip==24.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
python-mecab-ko==1.3.5
python-mecab-ko-dic==2.1.1.post2
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.11.0
typing-inspect==0.9.0
urllib3==2.2.1
wheel==0.43.0
whisper==1.1.10
yarl==1.9.4
